Move chainSetup from global magic to explicit snapshotSetup()
