+++ src/contracts/contract.test.ts
@@ -1,8 +1,10 @@
 import { getContractAddress } from "./contract";
 import { ethers } from "ethers";
 import { setupConfig } from "../test/sdkTestConfig";
+import { snapshotSetup } from "../test/hardhatRPC";
 
 describe("Contracts", () => {
+  snapshotSetup();
   let provider: ethers.providers.JsonRpcProvider;
   beforeAll(async () => {
     ({ provider } = setupConfig());
