+++ src/social/handles.test.ts
@@ -2,7 +2,7 @@ import { isAvailable, availabilityFilter } from "./handles";
 import * as registry from "../contracts/registry";
 import { createCloneProxy } from "../contracts/identity";
 import { setupConfig } from "../test/sdkTestConfig";
-import { revertHardhat, snapshotHardhat } from "../test/hardhatRPC";
+import { revertHardhat, snapshotHardhat, snapshotSetup } from "../test/hardhatRPC";
 import { ethers } from "ethers";
 
 const createIdentityContract = async () => {
@@ -12,6 +12,8 @@ const createIdentityContract = async () => {
 };
 
 describe("handles", () => {
+  snapshotSetup();
+
   const notTakens = ["not-taken", "not-taken1", "not-taken2"];
   const takens = ["taken", "taken1", "taken2"];
